python 2.5 compatibility
